90 introduce FileSystemOutputProviderTests
